requirements.txt
deepspeed>=0.12.0
transformers>=4.43.0
accelerate>=0.32.0
torchvision>=0.17.0
torchaudio>=2.2.0
opencv-python-headless>=4.8.0
Pillow>=10.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
vllm>=0.4.0
pyyaml>=6.0.0
numpy>=1.24.0
